requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
PyMySQL==1.1.1
python-dotenv==1.1.0
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0
sqlalchemy==2.0.40
wtforms==3.2.1 